set up hairstyles, and events page
added plus svg to subnav heading for adding a hairstyle, event, or dancer

moved 'SelectionChoices' styled component to styles folder, because it is shared among forms.  It displays the choices that have been made in multi-selction options with  an 'x' button that removes the choice from the list of selections

Added class of 'selectionWindow to NewStudioMain styledComponent.  This class will hold all styles that are common between all selection windows.  At this time, the only styles that are not common are the grid column placements, because classes is in the first 10 columns, with the filters in last 2.